requirements.txt
jaxlib
jax
numpy
scipy
matplotlib
pyhdf
xlrd
pyyaml
mlflow
boto3
flatten-dict
flatdict
optax
tqdm
jaxopt
xarray
mlflow_export_import
